Fix undefined integer left shift
UBSan complains that `left shift of 255 by 24 places cannot be represented in type 'int'`.

+++ src/bitmap.c
@@ -200,7 +200,7 @@ gdip_property_get_long(int offset, void *value, guint32 *result)
 {
 	BYTE *ptr = (BYTE*)value;
 
-	*result = ptr[0] + (ptr[1] << 8) + (ptr[2] << 16) + (ptr[3] << 24);
+	*result = ptr[0] + (ptr[1] << 8) + (ptr[2] << 16) + ((guint32)ptr[3] << 24);
 
 	return Ok;
 }
@@ -221,8 +221,8 @@ gdip_property_get_rational(int offset, void *value, guint32 *numerator, guint32 
 {
 	BYTE *ptr = (BYTE*)value;
 
-	*numerator = ptr[0] + (ptr[1] << 8) + (ptr[2] << 16) + (ptr[3] << 24);
-	*denominator = ptr[4] + (ptr[5] << 8) + (ptr[6] << 16) + (ptr[7] << 24);
+	*numerator = ptr[0] + (ptr[1] << 8) + (ptr[2] << 16) + ((guint32)ptr[3] << 24);
+	*denominator = ptr[4] + (ptr[5] << 8) + (ptr[6] << 16) + ((guint32)ptr[7] << 24);
 
 	return Ok;
 }
@@ -1537,7 +1537,7 @@ gdip_pixel_stream_get_next (StreamingState *state)
 		 */
 		if (state->pixels_per_byte == -4) {
 #if WORDS_BIGENDIAN
-			ret = state->scan [0] | (state->scan [1] << 8) | (state->scan [2] << 16) | (state->scan [3] << 24);
+			ret = state->scan [0] | (state->scan [1] << 8) | (state->scan [2] << 16) | ((guint32)state->scan [3] << 24);
 #else
 			ret = *(unsigned int *)state->scan;
 #endif
